+++ lion_master/src/app/layout/rdf-modelling/modelling-container/modelling-container.module.ts
@@ -7,14 +7,14 @@ import { FormsModule } from '@angular/forms';
 import { ReactiveFormsModule } from '@angular/forms';
 
 // rdf model components
-import { DashboardComponent } from './modelling-components/dashboard/dashboard.component';
-import { VDI3682Component } from './modelling-components/vdi3682/vdi3682.component';
-import { Dinen61360Component } from './modelling-components/dinen61360/dinen61360.component';
-import { Vdi2206Component } from './modelling-components/vdi2206/vdi2206.component';
-import { WadlComponent } from './modelling-components/wadl/wadl.component';
-import { Isa88Component } from './modelling-components/isa88/isa88.component';
-import { Iso22400_2Component } from './modelling-components/iso22400-2/iso22400-2.component'
-import { ConfigurationsComponent } from './modelling-components/configurations/configurations.component';
+// import { DashboardComponent } from './modelling-components/dashboard/dashboard.component';
+// import { VDI3682Component } from '../../../modelling/vdi3682/vdi3682.component';
+// import { Dinen61360Component } from './modelling-components/dinen61360/dinen61360.component';
+// import { Vdi2206Component } from './modelling-components/vdi2206/vdi2206.component';
+// import { WadlComponent } from '../../../modelling/wadl/wadl.component';
+// import { Isa88Component } from './modelling-components/isa88/isa88.component';
+// import { Iso22400_2Component } from './modelling-components/iso22400-2/iso22400-2.component'
+// import { ConfigurationsComponent } from './modelling-components/configurations/configurations.component';
 
 
 
@@ -35,14 +35,14 @@ import { ChartsModule as Ng2Charts } from 'ng2-charts';
   declarations: [
     ModellingContainerComponent,
     // rdf model components
-    DashboardComponent,
-    VDI3682Component,
-    Dinen61360Component,
-    Vdi2206Component,
-    WadlComponent,
-    Isa88Component,
-    Iso22400_2Component,
-    ConfigurationsComponent,
+    // DashboardComponent,
+    // VDI3682Component,
+    // Dinen61360Component,
+    // Vdi2206Component,
+    // WadlComponent,
+    // Isa88Component,
+    // Iso22400_2Component,
+    // ConfigurationsComponent,
     
     // util components
     TableComponent
