+++ lion_master/src/app/layout/tables/table/table.module.ts
@@ -1,0 +1,21 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { NgbCarouselModule, NgbAlertModule } from '@ng-bootstrap/ng-bootstrap';
+import { FormsModule } from '@angular/forms';
+import { ReactiveFormsModule } from '@angular/forms';
+
+import { TableComponent } from './table.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    NgbCarouselModule,
+    NgbAlertModule,
+    FormsModule,
+    ReactiveFormsModule
+  ],
+  declarations: [TableComponent],
+  exports: [TableComponent]
+})
+export class TableModule { }
