+++ lion_master/src/app/modelling/dinen61360/dinen61360.component.ts
@@ -3,25 +3,25 @@ import { FormBuilder, FormGroup } from '@angular/forms';
 import { FormArray } from '@angular/forms';
 import { Validators } from '@angular/forms';
 
-import { Dinen61360Service, DINEN61360Variables } from '../../../rdf-models/dinen61360Model.service';
-import { Isa88ModelService } from '../../../rdf-models/isa88Model.service';
-import { Vdi3682ModelService } from '../../../rdf-models/vdi3682Model.service';
-import { Vdi2206ModelService } from '../../../rdf-models/vdi2206Model.service';
-import { Iso22400_2ModelService } from '../../../rdf-models/iso22400_2Model.service';
+import { Dinen61360Service, DINEN61360Variables } from '../../layout/rdf-modelling/rdf-models/dinen61360Model.service';
+import { Isa88ModelService } from '../../layout/rdf-modelling/rdf-models/isa88Model.service';
+import { Vdi3682ModelService } from '../../layout/rdf-modelling/rdf-models/vdi3682Model.service';
+import { Vdi2206ModelService } from '../../layout/rdf-modelling/rdf-models/vdi2206Model.service';
+import { Iso22400_2ModelService } from '../../layout/rdf-modelling/rdf-models/iso22400_2Model.service';
 
-import { EclassService } from '../../../../../shared/services/backEnd/eclass.service';
-import { PrefixesService } from '../../../../../shared/services/prefixes.service';
+import { EclassService } from '../../shared/services/backEnd/eclass.service';
+import { PrefixesService } from '../../shared/services/prefixes.service';
 
-import { DataLoaderService } from '../../../../../shared/services/dataLoader.service';
-import { MessagesService } from '../../../../../shared/services/messages.service';
+import { DataLoaderService } from '../../shared/services/dataLoader.service';
+import { MessagesService } from '../../shared/services/messages.service';
 import { take } from 'rxjs/operators';
-import { Tables } from '../../../utils/tables';
+import { Tables } from '../../layout/rdf-modelling/utils/tables';
 
 
 @Component({
   selector: 'app-dinen61360',
   templateUrl: './dinen61360.component.html',
-  styleUrls: ['../../../../../../app/app.component.scss', './dinen61360.component.scss']
+  styleUrls: ['../../app.component.scss', './dinen61360.component.scss']
 })
 export class Dinen61360Component implements OnInit {
   // util variables
