+++ lion_master/src/app/modelling/modelling.component.html
@@ -1,6 +1,4 @@
 <app-sidebar (collapsedEvent)="receiveCollapsed($event)"></app-sidebar>
 <section class="main-container" [ngClass]="{collapsed: collapedSideBar}">
-  <p>asdasdasd</p>
-
-    <!-- <router-outlet></router-outlet> -->
+    <router-outlet></router-outlet>
 </section>
