+++ lion_master/src/app/modelling/sidebar/sidebar.component.html
@@ -1,26 +1,26 @@
 <nav class="sidebar" [ngClass]="{sidebarPushRight: isActive, collapsed: collapsed}">
     <div class="list-group">
-        <a [routerLink]="['model/vdi3682']" [routerLinkActive]="['router-link-active']" class="list-group-item">
+        <a [routerLink]="['vdi3682']" [routerLinkActive]="['router-link-active']" class="list-group-item">
             <i class="fas fa-anchor"></i>&nbsp;
             <span> VDI 3682 Template </span>
         </a>
-        <a [routerLink]="['model/dinen61360']" [routerLinkActive]="['router-link-active']" class="list-group-item">
+        <a [routerLink]="['dinen61360']" [routerLinkActive]="['router-link-active']" class="list-group-item">
             <i class="fas fa-database"></i>&nbsp;
             <span> DIN EN 61360 Template </span>
         </a>
-        <a [routerLink]="['model/vdi2206']" [routerLinkActive]="['router-link-active']" class="list-group-item">
+        <a [routerLink]="['vdi2206']" [routerLinkActive]="['router-link-active']" class="list-group-item">
             <i class="fas fa-sitemap"></i>&nbsp;
             <span> VDI2206 Template </span>
         </a>
-        <a [routerLink]="['model/wadl']" [routerLinkActive]="['router-link-active']" class="list-group-item">
+        <a [routerLink]="['wadl']" [routerLinkActive]="['router-link-active']" class="list-group-item">
             <i class="fas fa-globe"></i>&nbsp;
             <span> WADL Template </span>
         </a>
-        <a [routerLink]="['model/isa88']" [routerLinkActive]="['router-link-active']" class="list-group-item">
+        <a [routerLink]="['isa88']" [routerLinkActive]="['router-link-active']" class="list-group-item">
             <i class="fas fa-project-diagram"></i>&nbsp;
             <span> ISA 88 Template </span>
         </a>
-        <a [routerLink]="['model/iso22400-2']" [routerLinkActive]="['router-link-active']" class="list-group-item">
+        <a [routerLink]="['iso22400-2']" [routerLinkActive]="['router-link-active']" class="list-group-item">
             <i class="fab fa-keycdn"></i>&nbsp;
             <span> ISO 22400-2 Template </span>
         </a>
