+++ lion_master/src/app/modelling/wadl/assets/ontologicalDataType.png
